Opal panel: moved control groups up to make room for new stuff.
The FencePost class now allows calculation of a coordinate
from a floating point index, not just an integer index.
This allowed me to be more flexible with placement of items
whose panel heights are different.

In Opal, I used fractional index to help put the control
group knobs exactly where I wanted them.

+++ util/sapphire.py
@@ -138,7 +138,7 @@ class FencePost:
         self.nItems = nItems
         self.delta = (highValue - lowValue) / (nItems - 1)
 
-    def value(self, itemIndex: int) -> float:
+    def value(self, itemIndex: float) -> float:
         return self.lowValue + (itemIndex * self.delta)
